Add files via upload
In this file, you will calculate Pi (π) is to use the Gregory-Leibniz Series.

question is : 
If you continued this pattern forever you would be able to calculate ∏ exactly. If however you start to add up the first few terms, you will begin to get an approximation for Pi (π). The problem with the series above is that you need to add up a lot of terms in order to get an accurate approximation of Pi (π). Your task is to write a program that will take an integer input

(n)	and calculate Pi (π) value using the Gregory-Leibniz Series. The output of the program will be the difference of generated Pi (π) from standard value of Pi = 3.14.

Sample Inputs:

Enter an integer value (n): 0
Sample Outputs:
 0.86
Sample Inputs:

Enter an integer value (n): 3
Sample Outputs:

0.2447